Question Wall
The Question wall will show a question with one correct answer and three wrong answers. These answers will be randomizes, minus the last one, so the the answer order is not always the same. this prevents people from memorizing the order of correct answers and makes them memorize the correct answer itself.