Search and Highlight + Rail System still Works
For each Group of Hazards, they should be linked through a list on a GameObject with the SearchAndFindManager. This will track when they have been idintified and will highlight those that have not been. There is a function to check and see if all have been cleared, if not it highlights those missing. If so, an event pops off. Also tested to see if the rail system still works and it does. All it neede were points, a blindfold, and a player to move.